+++ R/meteo_imgw_daily.R
@@ -90,6 +90,7 @@ meteo_imgw_daily_bp = function(rank,
 
   # match WMO ID of a given station(s) to download selectively for SYNOP stations
   if (!is.null(station) && rank == "synop") {
+    station = find_all_station_names(station)
     ids_to_download = match_imgw_wmoid_inds(station)
   } else {
     ids_to_download = NULL