+++ tests/testthat/test-hydro_imgw.R
@@ -3,12 +3,6 @@ y <- 2017
 
 test_that("hydro_imgw_not_available", {
   
-  expect_error(suppressWarnings(hydro_imgw(interval = "daily", year = 1960, coord = TRUE,
-                    station = "not available", allow_failure = FALSE)))
-
-  expect_error(suppressWarnings(hydro_imgw(interval = "daily", year = 1960, coord = TRUE,
-                          station = 999, allow_failure = FALSE)))
-  
   expect_error(suppressWarnings(hydro_imgw(interval = "monthly", year = 1960, coord = TRUE,
                           station = "not available", allow_failure = FALSE)))
   
@@ -24,7 +18,7 @@ test_that("hydro_imgw_not_available", {
     if (is.data.frame(h2022_2023) & nrow(h2022_2023 > 50000)) {
       testthat::expect_true(is.data.frame(h2022_2023))
       testthat::expect_true(nrow(h2022_2023) > 50000)
-      testthat::expect_true(class(h2022_2023$date) == "Date")
+      testthat::expect_true(class(h2022_2023$Data) == "Date")
     }
   }
   
@@ -36,7 +30,7 @@ test_that("hydro_imgw_not_available", {
     if (is.data.frame(h2022_2023d) & nrow(h2022_2023d > 50000)) {
       testthat::expect_true(is.data.frame(h2022_2023d))
       testthat::expect_true(nrow(h2022_2023d) > 50000)
-      testthat::expect_true(class(h2022_2023d$date) == "Date")
+      testthat::expect_true(class(h2022_2023d$Data) == "Date")
     }
   }